feat: Implement Avatar Publisher functionality
- Added PublisherInstance class for managing avatar publishing UI and logic.
- Introduced PublisherPanel class to initialize and manage the publisher panel.
- Created PublisherThumbnail class for handling thumbnail uploads and previews.
- Developed PublisherUI class to manage display states and UI updates.
- Updated AvatarDescriptor to use ushort for publishVersion.
- Added new UXML files for publisher panel layout and structure.
- Enhanced language assets with success message for avatar build completion.
- Modified builder panel layout for improved user experience.
- Updated mod configuration to include PublisherPanel in the editor.

+++ Packages/api.nox.avatar/cck/AvatarDescriptor.cs
@@ -18,7 +18,7 @@ namespace Nox.CCK.Avatars {
 		public Platform target;
 		public uint     publishId;
 		public string   publishServer;
-		public uint     publishVersion;
+		public ushort   publishVersion;
 		#endif
 
 		#endregion
